Cast flip matrix to int

+++ preprocessing/dem/Shop_01.cpp
@@ -113,7 +113,7 @@ Matrix3r Shop::flipCell(const Matrix3r& _flip)
 	cell->hSize = new_hSize;
 
 	flip = _flip; // Just so -Werror=unused-parameter doesn't bother me for now (to be improved)
-	flip = cell->hSize.inverse() * new_hSize -  Matrix3r::Identity();
+	flip = (cell->hSize.inverse() * new_hSize -  Matrix3r::Identity()).cast<int>();
 
 	cell->postLoad(*cell);
